Refactor PhysicalDeviceSelector and check the Vulkan API version supported by candidate devices.
Currently, the PhysicalDeviceSelector (and thus also SimpleDeviceSelector) is responsible for:
- Checking which physical devices support the minimum requirements of the application.
- From those physical devices, select the 'best' one.

This commit shifts the first responsibility to a fresh class BasicDeviceFilter.
Unfortunately, this is a breaking change since it breaks any program that uses
a custom PhysicalDeviceSelector, but I believe the design is a lot nicer.

Unlike the old SimpleDeviceSelector, the new BasicDeviceFilter automatically
filters out devices that don't support the requested Vulkan API version.