Only use acquire fences when needed
Currently, we always assign both an acquire fence and a present fence to
each swapchain image, which is not needed:
- with swapchain maintenance, only the present fences are needed
- without swapchain maintenance, only the acquire fences are needed

Furthermore, this commit prevents old swapchains from piling up while
the window is being resized quickly.

Both changes should reduce the number of fences and semaphores that
are created. The best practices still complain that we create too
many, but I don't believe that a limit of 3 fences and 10 semaphores
is reasonable.